v0.13
Adding new parameters (no_title, no_description, no_example and minimal_schema) and fleshing out unit tests to support them.

minimal_schema turns on all the no_* features to try and generate as minimal of a schema as possible.

The unit tests will populate the data file (if you make changes to the data structure and want to keep things in sync) and the schemas generated are now compared against what was on disk. We might need to make that be canonical so that if different versions of perl generate the json differently we mask that.

You need to set the environment variable GENERATE_SCHEMA_FILES=1 when calling make test so the data files are generated.

Ex: GENERATE_SCHEMA_FILES=1 make test

Otherwise the computed schema will be compared against what was provided in the t/schemas/ directory and differences will be alerted on.